requirements.txt
# Core
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0
click>=8.0.0

# Audio
pyaudio>=0.2.11
numpy>=1.20.0

# Speech-to-Text
websockets>=10.0
deepgram-sdk>=3.0.0

# Command Matching
fuzzywuzzy>=0.18.0

# macOS Automation
PyObjC>=9.0

# Logging & Utils
colorama>=0.4.0

